# docs/swagger.yaml
definitions:
  dto.CreateProductRequest:
    properties:
      price:
        example: 50
        type: number
      product_name:
        example: Keyboard
        type: string
    type: object
  dto.Response:
    properties:
      message:
        type: string
    type: object
  dto.UpdateProductRequest:
    properties:
      price:
        example: 50
        type: number
      product_name:
        example: Keyboard
        type: string
    type: object
  model.Product:
    properties:
      id_product:
        type: integer
      price:
        type: number
      product_name:
        type: string
    type: object
info:
  contact: {}
paths:
  /product:
    post:
      consumes:
      - application/json
      description: Create a new product in the database
      parameters:
      - description: Product to create
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Create a new product
      tags:
      - products
  /product/{productID}:
    delete:
      description: Delete a product from the database
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete a product
      tags:
      - products
    get:
      description: Retrieve a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product in the database
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      - description: Product data to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update an existing product
      tags:
      - products
  /products:
    get:
      description: Get all products from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List all products
      tags:
      - products
swagger: "2.0"